Finding the best hospital in a state
The function reads the outcome-of-care-measures.csv le and returns a
character vector
with the name of the hospital that has the best (i.e. lowest) 30-day
mortality for the specied outcome
in that state. The hospital name is the name provided in the
Hospital.Name variable. The outcomes can
be one of \heart attack", \heart failure", or \pneumonia". Hospitals
that do not have data on a particular
outcome should be excluded from the set of hospitals when deciding the
rankings.
Handling ties. If there is a tie for the best hospital for a given
outcome, then the hospital names should
be sorted in alphabetical order and the rst hospital in that set
should be chosen (i.e. if hospitals \b", \c",
and \f" are tied for best, then hospital \b" should be returned).