Ranking hospitals by outcome in a state
The function reads the outcome-of-care-measures.csv le and returns a
character vector with the name
of the hospital that has the ranking specied by the num argument. F
The num argument can take values \best", \worst", or an integer
indicating the ranking
(smaller numbers are better). If the number given by num is larger than
the number of hospitals in that
state, then the function should return NA. Hospitals that do not have
data on a particular outcome should
be excluded from the set of hospitals when deciding the rankings.